# Conflicts:
#	src/main/java/org/ljx/dao/WarehouseRecordMapper.java
#	src/main/java/org/ljx/entity/WarehouseRecord.java
完成商品管理基础功能

+++ src/main/java/org/ljx/service/warehouse/WarehouseService.java
@@ -11,7 +11,7 @@ import java.util.List;
  */
 public interface WarehouseService {
 
-    public List<Warehouse> list(byte storeId, byte status);
+    public List<Warehouse> list(int storeId, byte status);
 
     public void insert(Warehouse warehouse);
